WIP: tracking down errors (see valgrind reports when tests are run)
Problem: We delete a queue while some workers are still listening on it.
Solutions:
- Maybe queues should have a list of listeners/workers so that we wait
  for a listener to stop listening before nuking the queue.
- Maybe make it a callers problem - they should issue a kill-all and
  wait-all before proceeding to nuke everything.
- Additionally, lib_destroy() does the kill-all/wait-all sequence
  before removing a queue.

Too many things to check here - after all, maybe a listener is listening
on a queue without registering as a consumer of that queue (allowed). We
could ana a semaphore tot he queue structure, and wait for that to go to
zero (while not allowing new listeners).