BUGFIX: double-free/memleaks in worker_del on lib_destroy()
Fix involved walking through the container queue and manually sending
SIGNAL_TERMINAT, then waiting for termination of each worker. The is
still the problem that the container is not locked during this process
and that any new workers created will be memleaked as teh library is
destroyed (or when the program ends).

As an alternative, perhaps the worker_run thread function should kill
workers who experience more than 5 failures per second?