Lots of changes to the application for it to work under Windows.
Under Windows the scheduling is different (different algorithms,
different quanta, etc) which revealed some bugs in the program.

These were logic bugs, and it turned out that there was no need
send kill signals to workers. Now the application simply waits for
queue to be empty and then exits.

If the user forces an exit with CTRL-c that works too, but the messages
already in the queue are discarded (valgrind might report them as memory
leaks).